Fix type of local variable, fixing a 64-to-32 warning.

+++ Core/Source/GrowlGNTPHeaderItem.m
@@ -63,7 +63,7 @@
 #define HEADER_DELIMITER_LENGTH 2
 #define CRLF_LENGTH 2
 		
-		int endOfHeaderName = [headerLine rangeOfString:HEADER_DELIMITER options:NSLiteralSearch].location;
+		NSInteger endOfHeaderName = [headerLine rangeOfString:HEADER_DELIMITER options:NSLiteralSearch].location;
 		if (endOfHeaderName == NSNotFound) {
 			/* Malformed header; no "name: value" setup */
 			if (outError)
